rename reader dtype->tod_dtype

+++ src/furax/mapmaking/mapmaker.py
@@ -132,7 +132,7 @@ class MultiObservationMapMaker(Generic[T]):
             requested_fields=data_field_names,
             demodulated=self.config.demodulated,
             stokes=self.config.stokes,
-            dtype=self.config.dtype,
+            tod_dtype=self.config.dtype,
         )
 
     def make_maps(self) -> MapMakingResults:
